+++ app/routes/_auth+/onboarding.company.tsx
@@ -1,0 +1,7 @@
+export const action = () => {
+   return redirectWithToast(
+		safeRedirect(redirectTo),
+		{ title: 'Welcome', description: 'Thanks for signing up!' },
+		{ headers }, 
+		)
+}
